Extract `Person` class to a separate file
